Feat: 모델연동 완료 chat 반환 완료 세션별 레포트 진행중

See merge request s12-final/S12P31S202!64

+++ backend/damdam/src/main/java/com/ssafy/damdam/domain/users/service/UserServiceImpl.java
@@ -141,8 +141,7 @@ public class UserServiceImpl implements UserService {
 		Users user = validateUser();
 
 		UserSurvey survey = userSurveyRepository
-				.findByUsers_UserId(user.getUserId())
-				.orElseThrow(() -> new UserException(USER_SURVEY_NOT_FOUND));
+				.findByUsers_UserId(user.getUserId()).orElse(null);
 
 		return UserSurveyOutputDto.fromEntity(survey);
 	}
@@ -160,9 +159,11 @@ public class UserServiceImpl implements UserService {
 	public void deleteSurvey() {
 		Users user = validateUser();
 		UserSurvey survey = userSurveyRepository
-			.findByUsers_UserId(user.getUserId())
-			.orElseThrow(() -> new UserException(USER_SURVEY_NOT_FOUND));
+			.findByUsers_UserId(user.getUserId()).orElse(null);
 
+		if (survey == null) {
+			throw new UserException(USER_SURVEY_NOT_FOUND);
+		}
 		userSurveyRepository.delete(survey);
 	}
 }
